chore: move GuUserData into it's own file
