+++ src/constructs/autoscaling/asg.ts
@@ -1,6 +1,6 @@
 import type { AutoScalingGroupProps, CfnAutoScalingGroup } from "@aws-cdk/aws-autoscaling";
 import { AutoScalingGroup } from "@aws-cdk/aws-autoscaling";
-import type { ISecurityGroup, MachineImage, MachineImageConfig, S3DownloadOptions } from "@aws-cdk/aws-ec2";
+import type { ISecurityGroup, MachineImage, MachineImageConfig } from "@aws-cdk/aws-ec2";
 import { InstanceType, OperatingSystemType, UserData } from "@aws-cdk/aws-ec2";
 import type { ApplicationTargetGroup } from "@aws-cdk/aws-elasticloadbalancingv2";
 import { Stage } from "../../constants";
@@ -32,24 +32,6 @@ export interface GuAutoScalingGroupProps
   additionalSecurityGroups?: ISecurityGroup[];
   targetGroup?: ApplicationTargetGroup;
   overrideId?: boolean;
-}
-
-export class GuUserData {
-  private _userData = UserData.forLinux();
-
-  get userData(): UserData {
-    return this._userData;
-  }
-
-  addCommands(...commands: string[]): GuUserData {
-    this._userData.addCommands(...commands);
-    return this;
-  }
-
-  addS3DownloadCommand(params: S3DownloadOptions): GuUserData {
-    this._userData.addS3DownloadCommand(params);
-    return this;
-  }
 }
 
 type GuStageDependentAsgProps = Record<Stage, GuAsgCapacityProps>;
