+++ src/constructs/autoscaling/index.ts
@@ -1,1 +1,2 @@
 export * from "./asg";
+export * from "./user-data";
