+++ src/constructs/autoscaling/user-data.ts
@@ -1,0 +1,20 @@
+import type { S3DownloadOptions } from "@aws-cdk/aws-ec2";
+import { UserData } from "@aws-cdk/aws-ec2";
+
+export class GuUserData {
+  private _userData = UserData.forLinux();
+
+  get userData(): UserData {
+    return this._userData;
+  }
+
+  addCommands(...commands: string[]): GuUserData {
+    this._userData.addCommands(...commands);
+    return this;
+  }
+
+  addS3DownloadCommand(params: S3DownloadOptions): GuUserData {
+    this._userData.addS3DownloadCommand(params);
+    return this;
+  }
+}
